Refactor Category entity and related components to use Long as the identifier type instead of Integer. This change includes updates to the Category class, CategoryController, CategoryService, and associated tests to ensure consistency across the codebase. Additionally, enhance validation constraints for category name and description fields.

+++ src/main/java/com/sigrap/category/CategoryInfo.java
@@ -1,8 +1,7 @@
 package com.sigrap.category;
 
-import java.time.LocalDateTime;
-
 import io.swagger.v3.oas.annotations.media.Schema;
+import java.time.LocalDateTime;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -59,7 +58,7 @@ public class CategoryInfo {
    * </ul></p>
    */
   @Schema(description = "Unique identifier of the category", example = "1")
-  private Integer id;
+  private Long id;
 
   /**
    * The name of the category.
